Remove postPatch for Dual Architecture Support
This commit removes the 'postPatch' step, which is no longer necessary due to updated compatibility and requirements across x86 and AArch64 systems. Previously, the patch was needed to align with architecture-specific configurations, but subsequent changes in the base configurations of the package now offer native support for both architectures without this step.